Haut_parleur sort sur les 2 côtés.

+++ S5/APP5/S5e_APP6/source/c/Audio_driver.c
@@ -25,7 +25,7 @@
 
 extern far void vectors();   // Vecteurs d'interruption
 
-extern short output;
+extern int output;
 extern unsigned int input;
 extern int reception_micro;
 /****************************************************************************
